# src/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WiserLP API Docs",
    "version": "v1"
  },
  "tags": [{
    "name": "Client Endpoints"
  }],
  "paths": {
    "/register-client": {
      "post": {
        "tags": [
          "Client Endpoints"
        ],
        "operationId": "register-client",
        "summary": "Register a new client, you can do so using both username/password or appkey/secret/accessToken/accessTokenSecret, consult the models below",
        "parameters": [
          {
            "name": "User",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User - Authentication with Password"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Register success"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/unregister-client": {
      "delete": {
        "tags": [
          "Client Endpoints"
        ],
        "operationId": "unregister-client",
        "summary": "Remove entirely from the system a given user",
        "parameters": [
          {
            "name": "accountId",
            "in": "body",
            "example": "{\n  \"accountId\": \"ABCDE-12345-MNBVC-98765\"\n}",
            "required": true
          }
        ]
      }
    },
    "/update-metadata": {
      "put": {
        "tags": [
          "Client Endpoints"
        ],
        "operationId": "update-metadata",
        "summary": "Update all of the account-related parameters, except for the username",
        "parameters": [
          {
            "name": "User",
            "in": "body",
            "example": "{\n  \"accountId\": \"ABCDE-12345-MNBVC-98765\",\n  \"webhooks\": {\n    \"new_conversation_webhook\": \"https://hooks.zapier.com/hooks/catch/1646904/73834/\",\n    \"new_file_in_conversation_webhook\": \"https://hooks.zapier.com/hooks/catch/1234443/234234/\",\n    \"new_message_arrived_webhook\": \"https://hooks.zapier.com/hooks/catch/1234443/293734/\"\n  }\n}",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Account updated"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/send-message": {
      "post": {
        "tags": [
          "Actions"
        ],
        "operationId": "send-message",
        "summary": "Send a message to any open conversation",
        "parameters": [
          {
            "name": "Message",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message sent"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/conversation-details": {
      "get": {
        "tags": [
          "Client Endpoints"
        ],
        "operationId": "conversation-details",
        "summary": "Retrieve conversation details, passing a specific agentId and conversationId",
        "parameters": [
          {
            "name": "accountId",
            "in": "body",
            "example": "{\n  \"accountId\": \"ABCDE-12345-MNBVC-98765\"\n}",
            "required": true
          },
          {
            "name": "convId",
            "in": "body",
            "example": "{\n  \"convId\": \"123456789\"\n}",
            "required": true
          }
        ]
      }
    }
  },
  "definitions": {
    "User - Authentication with Secret": {
      "required": [
        "username",
        "liveperson_accountid",
        "liveperson_appkey",
        "liveperson_secret",
        "liveperson_accesstoken",
        "liveperson_accesstokensecret"
      ],
      "properties": {
        "username": {
          "type": "string"
        },
        "liveperson_password": {
          "type": "string"
        },
        "liveperson_accountid": {
          "type": "string"
        },
        "liveperson_appkey": {
          "type": "string"
        },
        "liveperson_secret": {
          "type": "string"
        },
        "liveperson_accesstoken": {
          "type": "string"
        },
		    "liveperson_accesstokensecret": {
          "type": "string"
        },
        "webhooks": {
          "$ref": "#/definitions/Webhooks"
        }
      },
      "example": {
        "username": "John Doe",
        "liveperson_accountid": "1234567890",
        "liveperson_appkey": "1234567890",
        "liveperson_secret": "1234567890",
        "liveperson_accesstoken": "1234567890",
        "liveperson_accesstokensecret": "1234567890",
        "webhooks": {
          "new_conversation_webhook": "https://hooks.zapier.com/hooks/catch/1646904/73834/",
          "new_file_in_conversation_webhook": "https://hooks.zapier.com/hooks/catch/1234443/234234/"
        }
      }
    },
    "User - Authentication with Password": {
      "required": [
        "username",
        "liveperson_accountid",
        "liveperson_password"
      ],
      "properties": {
        "username": {
          "type": "string"
        },
        "liveperson_password": {
          "type": "string"
        },
        "liveperson_accountid": {
          "type": "string"
        },
        "webhooks": {
          "$ref": "#/definitions/Webhooks"
        }
      },
      "example": {
        "username": "John Doe",
        "liveperson_accountid": "1234567890",
        "liveperson_password": "1234567890",
        "webhooks": {
          "new_conversation_webhook": "https://hooks.zapier.com/hooks/catch/1646904/73834/",
          "new_file_in_conversation_webhook": "https://hooks.zapier.com/hooks/catch/1234443/234234/",
          "new_message_arrived_webhook": "https://hooks.zapier.com/hooks/catch/1234443/293734/"
        }
      }
    },
    "Message": {
      "type": "object",
      "required": [
        "credentials",
        "message"
      ],
      "properties": {
        "credentials": {
          "$ref": "#/definitions/User - Authentication with Password"
        },
        "message": {
          "type": "object",
          "required": [
            "dialogId",
            "contentType",
            "message"
          ],
          "properties": {
            "dialogId": {
              "type": "string"
            },
            "contentType": {
              "type": "string"
            },
            "message": {
              "type": "string"
            }
          }
        }
      },
      "example": {
      	"credentials": {
      		"username": "john@doe.com",
      		"liveperson_accountid": "978654321",
      		"liveperson_password": "MySafePasswowrd1234"
      	},
      	"message": {
      		"dialogId": "ABCDE-12345-MNBVC-98765",
      		"contentType": "text/plain",
      		"message": "Hello!"
      	}
      }
    },
    "Webhooks": {
      "type": "object",
      "properties": {
        "new_conversation_webhook": {
          "type": "string"
        },
        "new_file_in_conversation_webhook": {
          "type": "string"
        },
        "new_message_arrived_webhook": {
          "type": "string"
        }
      }
    }
  }
}